Move the cluster list toolbar actions above the table
To align to OCM cluster list the botton toolbar is removed in clusters
list and actions are put into clusters table toolbar.
